+++ packages/ui/Transition/Transition.js
@@ -152,12 +152,12 @@ const TransitionBase = kind({
 			timingFunction && css[timingFunction],
 			css[type]
 		),
-		innerStyle: ({clipWidth, direction, type, visible}) => (type === 'clip' && (direction === 'left' || direction === 'right')) ? {
+		innerStyle: ({clipWidth, direction, type}) => (type === 'clip' && (direction === 'left' || direction === 'right')) ? {
 			width: clipWidth
 		} : null,
-		style: ({clipHeight, clipWidth, direction, type, visible, style}) => (type === 'clip') ? {
+		style: ({clipHeight, direction, type, visible, style}) => (type === 'clip') ? {
 			...style,
-			height: (visible && (direction === 'up' || direction === 'down')) ? clipHeight : null,
+			height: (visible && (direction === 'up' || direction === 'down')) ? clipHeight : (style && style.height), // use existing height if present and we didn't pass the earlier test
 			overflow: 'hidden'
 		} : style,
 		childRef: ({childRef, noAnimation, children}) => (noAnimation || !children) ? null : childRef
@@ -212,19 +212,9 @@ class Transition extends React.Component {
 		 *
 		 * @type {Number}
 		 * @default null
+		 * @deprecated Removed in 2.0.0 since it ignored anyway.
 		 * @public
 		 */
-		clipHeight: PropTypes.number,
-
-		/**
-		 * The width of the transition when `type` is set to `'clip'`, used when direction is 'left'
-		 * or 'right'.
-		 *
-		 * @type {Number}
-		 * @default null
-		 * @public
-		 */
-		clipWidth: PropTypes.number,
 
 		/**
 		 * The direction of transition (i.e. where the component will move *to*; the destination).
@@ -400,7 +390,7 @@ class Transition extends React.Component {
 		if (this.childNode) {
 			const initialHeight = this.childNode.scrollHeight;
 			const initialWidth = this.childNode.scrollWidth;
-			if (initialHeight !== this.state.initialHeight || initialHeight !== this.state.initialWidth) {
+			if (initialHeight !== this.state.initialHeight || initialWidth !== this.state.initialWidth) {
 				this.setState({
 					initialHeight,
 					initialWidth,